requirements.txt
python-telegram-bot==20.7
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.23
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
Pillow==10.0.1
WTForms==3.0.1
Flask-WTF==1.1.1